+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @workflow/cli
 
+## 4.2.8
+
+### Patch Changes
+
+- Updated dependencies [[`5655fcb`](https://github.com/vercel/workflow/commit/5655fcbd4d5ed45c6669666fef8c817a286eec33)]:
+  - @workflow/world-vercel@4.3.2
+  - @workflow/core@4.3.1
+  - @workflow/web@4.1.9
+  - @workflow/builders@4.0.9
+
+## 4.2.7
+
+### Patch Changes
+
+- Updated dependencies [[`5a0ce9a`](https://github.com/vercel/workflow/commit/5a0ce9a700bb433ead4676d5d7f6bdb0bc6a6fdd), [`1e32d05`](https://github.com/vercel/workflow/commit/1e32d05758d590134f04f82ea41ab9add96104d5), [`5fd7d9c`](https://github.com/vercel/workflow/commit/5fd7d9c2a1bca6e6ecaac7c0577e1ac58960711f), [`9092640`](https://github.com/vercel/workflow/commit/90926400131be4003e65d970deb5d67ae29778a2)]:
+  - @workflow/core@4.3.0
+  - @workflow/world@4.1.4
+  - @workflow/errors@4.1.3
+  - @workflow/builders@4.0.8
+  - @workflow/web@4.1.8
+  - @workflow/world-local@4.1.4
+  - @workflow/world-vercel@4.3.1
+
 ## 4.2.6
 
 ### Patch Changes