+++ packages/nest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/nest
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.9
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.8
+
 ## 0.0.6
 
 ### Patch Changes