+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/nuxt
 
+## 4.0.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/nitro@4.0.9
+
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/nitro@4.0.8
+
 ## 4.0.7
 
 ### Patch Changes