+++ packages/sveltekit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @workflow/sveltekit
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.9
+  - @workflow/rollup@4.0.8
+  - @workflow/vite@4.0.8
+
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.8
+  - @workflow/rollup@4.0.7
+  - @workflow/vite@4.0.7
+
 ## 4.0.6
 
 ### Patch Changes