+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/vite
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.9
+
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.8
+
 ## 4.0.6
 
 ### Patch Changes